+++ src/diagnostic/qbsclangcldiagnosticparser.ts
@@ -5,14 +5,22 @@ import { QbsToolchain } from '../protocol/qbsprotocolqbsmoduledata';
 import { substractOne } from '../qbsutils';
 
 export class QbsClangClDiagnosticParser extends QbsDiagnosticParser {
-    private readonly compilerRegexp = /^(.+\.\S+)\((\d+,\d+)\):\s(fatal error|error|warning):\s(.+)/;
+    private readonly compilerRegexp = /^(?<file>.+\.\S+)\((?<location>\d+,\d+)\):\s(?<severity>fatal error|error|warning):\s(?<details>.+)/;
+    private readonly linkerRegexp = /^(?:\d+>)?(.+?[^\s])\s?:\s(?<severity>(?:fatal )?error)\s(?<code>LNK\d+):\s(?<details>.+)$/;
 
     public constructor() { super(QbsToolchain.ClangCl); }
 
     protected parseLine(line: string): void {
+        if (this.parseCompilerMessage(line))
+            return;
+        else if (this.parseLinkerMessage(line))
+            return;
+    }
+
+    private parseCompilerMessage(line: string): boolean {
         const matches = this.compilerRegexp.exec(line);
         if (!matches)
-            return;
+            return false;
 
         const [, fsPath, location, severity, message] = matches;
         const range = (() => {
@@ -33,6 +41,27 @@ export class QbsClangClDiagnosticParser extends QbsDiagnosticParser {
         };
 
         this.insertDiagnostic(vscode.Uri.file(fsPath), diagnostic);
+        return true;
+    }
+
+    private parseLinkerMessage(line: string): boolean {
+        const matches = this.linkerRegexp.exec(line);
+        if (!matches)
+            return false;
+
+        const [, fsPath, severity, code, message] = matches;
+        // Fake location because the linker does not provide a positions or columns in an object files.
+        const range = new vscode.Range(0, 0, 0, 0);
+        const diagnostic: vscode.Diagnostic = {
+            source: this.toolchainType,
+            severity: QbsClangClDiagnosticParser.encodeSeverity(severity),
+            message,
+            range,
+            code
+        };
+
+        this.insertDiagnostic(vscode.Uri.file(fsPath), diagnostic);
+        return true;
     }
 
     private static encodeSeverity(severity: string): vscode.DiagnosticSeverity {